requirements.txt
zimply
python-baseconv
gensim
pandas
torch
torchvision
unidecode
spacy
openai
tqdm
nltk
bs4
scikit-learn
numpy
transformers 
torch 
datasets 
evaluate 
transformers[torch] 
evaluate
sentencepiece